+++ NEWS.md
@@ -48,13 +48,13 @@
 #### Changed
 * Added 65 antibiotics to the `antibiotics` data set, from the [Pharmaceuticals Community Register](http://ec.europa.eu/health/documents/community-register/html/atc.htm) of the European Commission
 * Removed columns `atc_group1_nl` and `atc_group2_nl` from the `antibiotics` data set
-* Function `atc_ddd` has been renamed `atc_online_ddd()`
-* Function `atc_groups` has been renamed `atc_online_groups()`
+* Functions `atc_ddd()` and `atc_groups()` have been renamed `atc_online_ddd()` and `atc_online_groups()`. The old function are deprecated and will be removed in a future version.
+* Function `guess_mo()` is now deprecated in favour of `as.mo()` and will be removed in future versions
+* Function `guess_atc()` is now deprecated in favour of `as.atc()` and will be removed in future versions
 * Function `eucast_rules()`:
   * Updated EUCAST Clinical breakpoints to [version 9.0 of 1 January 2019](http://www.eucast.org/clinical_breakpoints/)
   * Fixed a critical bug where some rules that depend on previous applied rules would not be applied adequately
   * Emphasised in manual that penicillin is meant as benzylpenicillin (ATC [J01CE01](https://www.whocc.no/atc_ddd_index/?code=J01CE01))
-* Function `guess_mo()` is now deprecated in favour of `as.mo()` and will be removed in future versions
 * Improvements for `as.mo()`:
   * Fix for vector containing only empty values
   * Finds better results when input is in other languages